fixed issue with params
