+++ Routes/index.js
@@ -7,7 +7,6 @@ router.get('/harian', require('./harian/index'));
 router.get('/harian/latest', require('./harian/latest'));
 router.get('/provinsi/:nama', require('./provinsi/index'));
 router.get('/provinsi/:nama/latest/', require('./provinsi/latest'))
-router.get('/provinsi/more', require('./provinsi/more'));
-router.get('/provinsi/alt', require('./provinsi/alt'));
+router.get('/provinsi/:nama/more/', require('./provinsi/more'));
 
 module.exports = router;
