+++ Routes/provinsi/index.js
@@ -1,9 +1,28 @@
 const { provData } = require('../../util/fetcher')
+const { provDataAlt } = require('../../util/fetcher')
 
 module.exports = async(req, res) => {
     const rawData = await provData()
-    console.log(req.params.nama)
-    if(req.params.nama){
+
+    if(req.params.nama == "alt"){
+        const datacollection = await provDataAlt();
+        const provData = datacollection.map(data => {
+            return data.attributes
+        })
+        provData.length = 34;
+        res.json(provData)
+    }else if(req.params.nama == "all"){
+        let datamodified = (rawData.list_data).filter(x => x.lokasi != null).map(data => {
+            return {
+                "provinsi": data.key,
+                "kasus": data.jumlah_kasus,
+                "dirawat": data.jumlah_dirawat,
+                "sembuh": data.jumlah_sembuh,
+                "meninggal": data.jumlah_meninggal
+            }
+        })
+        return res.json(datamodified);
+    }else{
         let datamodified = (rawData.list_data).filter(x => x.lokasi != null && x.key == req.params.nama.split('_').join(' ').toUpperCase()).map(data => {
             return {
                 "provinsi": data.key,
@@ -23,14 +42,5 @@ module.exports = async(req, res) => {
         return res.json(datamodified)
     }
 
-    let datamodified = (rawData.list_data).filter(x => x.lokasi != null).map(data => {
-        return {
-            "provinsi": data.key,
-            "kasus": data.jumlah_kasus,
-            "dirawat": data.jumlah_dirawat,
-            "sembuh": data.jumlah_sembuh,
-            "meninggal": data.jumlah_meninggal
-        }
-    })
-    return res.json(datamodified);
+    
 }
