+++ Routes/provinsi/more.js
@@ -2,8 +2,8 @@ const { provData } = require('../../util/fetcher')
 
 module.exports = async(req, res) => {
     const rawData = await provData();
-    if(req.query.name) {
-        let datamodified = (rawData.list_data).filter(x => x.lokasi != null && x.key == req.query.name.split('_').join(' ').toUpperCase()).map((data) => {
+    if(req.params.nama != "all") {
+        let datamodified = (rawData.list_data).filter(x => x.lokasi != null && x.key == req.params.nama.split('_').join(' ').toUpperCase()).map((data) => {
             const jeniskelamin = {}
             data['jenis_kelamin'].forEach(x => {
                 jeniskelamin[x['key'].toLowerCase()] = x['doc_count']
@@ -36,6 +36,7 @@ module.exports = async(req, res) => {
         });
         return res.json(datamodified);
     }
+    
     let datamodified = (rawData.list_data).filter(x => x.lokasi != null).map((data) => {
         const jeniskelamin = {}
         data['jenis_kelamin'].forEach(x => {
